# end/service/myrpc/proto/types.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "types.proto",
    "version": "version not set"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/types": {
      "get": {
        "summary": "设备类型列表",
        "operationId": "Types_typesGet",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/prototypesGetResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "fuzzy",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "Types"
        ]
      },
      "post": {
        "summary": "增加设备类型",
        "operationId": "Types_typesPost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/prototypesPostResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/prototypesPostReq"
            }
          }
        ],
        "tags": [
          "Types"
        ]
      },
      "put": {
        "summary": "修改设备类型",
        "operationId": "Types_typesPut",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/prototypesPutResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/prototypesPutReq"
            }
          }
        ],
        "tags": [
          "Types"
        ]
      }
    },
    "/v1/types/{id}": {
      "delete": {
        "summary": "删除设备类型",
        "operationId": "Types_typesDel",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/prototypesDelResp"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/runtimeError"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "tags": [
          "Types"
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "type_url": {
          "type": "string"
        },
        "value": {
          "type": "string",
          "format": "byte"
        }
      }
    },
    "prototypesDelResp": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "prototypesGetResp": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/prototypesInfo"
          }
        },
        "total": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "prototypesInfo": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int32"
        },
        "name": {
          "type": "string"
        },
        "details": {
          "type": "string"
        },
        "createtime": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "prototypesPostReq": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "details": {
          "type": "string"
        }
      }
    },
    "prototypesPostResp": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "prototypesPutReq": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "details": {
          "type": "string"
        },
        "id": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "prototypesPutResp": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "runtimeError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}